fix: use explicit i16 cast for libc flock constants
libc::F_WRLCK and F_UNLCK are i32 on Linux but i16 on macOS.
The cast is required for cross-platform builds but triggers
clippy::unnecessary_cast on macOS, so allow it with a comment.

+++ src/ida/lock.rs
@@ -266,11 +266,12 @@ pub(crate) fn detect_db_lock(path: &Path, _err: &IDAError) -> Option<String> {
 // Platform-specific file locking implementation
 
 #[cfg(unix)]
+#[allow(clippy::unnecessary_cast)] // F_WRLCK is i32 on Linux, i16 on macOS
 fn try_lock_file(file: &File) -> Result<(), u32> {
     use std::os::unix::io::AsRawFd;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -293,6 +294,7 @@ fn try_lock_file(_file: &File) -> Result<(), u32> {
 }
 
 #[cfg(unix)]
+#[allow(clippy::unnecessary_cast)] // F_WRLCK/F_UNLCK is i32 on Linux, i16 on macOS
 fn locked_by_pid(path: &Path) -> Option<u32> {
     use std::os::unix::io::AsRawFd;
 
@@ -304,7 +306,7 @@ fn locked_by_pid(path: &Path) -> Option<u32> {
         .ok()?;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -318,7 +320,7 @@ fn locked_by_pid(path: &Path) -> Option<u32> {
     if rc == -1 {
         return None;
     }
-    if fl.l_type == libc::F_UNLCK {
+    if fl.l_type == libc::F_UNLCK as i16 {
         None
     } else {
         Some(fl.l_pid as u32)
@@ -331,11 +333,12 @@ fn locked_by_pid(_path: &Path) -> Option<u32> {
 }
 
 #[cfg(unix)]
+#[allow(clippy::unnecessary_cast)] // F_WRLCK/F_UNLCK is i32 on Linux, i16 on macOS
 fn locked_by_pid_from_fd(file: &File) -> Option<u32> {
     use std::os::unix::io::AsRawFd;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -348,7 +351,7 @@ fn locked_by_pid_from_fd(file: &File) -> Option<u32> {
     if rc == -1 {
         return None;
     }
-    if fl.l_type == libc::F_UNLCK {
+    if fl.l_type == libc::F_UNLCK as i16 {
         None
     } else {
         Some(fl.l_pid as u32)
